Splits builds for prod Still needs updates to tests and docs
