+++ .bin/publish-npm.cjs
@@ -4,26 +4,29 @@ const spawn = require('cross-spawn')
 const semverValid = require('semver/functions/valid')
 const semverPrerelease = require('semver/functions/prerelease')
 
-const targetPath = 'dist'
+const rootTargetPath = 'dist'
 const args = process.argv.slice(2)
 
-const version = require(join(process.cwd(), targetPath, 'package.json')).version
+['core', 'admin', 'sso'].forEach(target => {
+  const targetPath = join(rootTargetPath, target)
+  const version = require(join(process.cwd(), targetPath, 'package.json')).version
 
-if (!semverValid(version)) {
-  console.error(`Error: invalid version "${version}"`)
-  process.exit(1)
-}
-
-const prerelease = semverPrerelease(version)
-const tag = (prerelease && prerelease.length) ? prerelease.filter(part => typeof part === 'string').join('.') || 'prerelease' : false
-const tagArgs = tag ? ['--tag', tag] : []
-const npm = /^win/.test(process.platform) ? 'npm.cmd' : 'npm'
-
-fs.writeFileSync(join(process.cwd(), targetPath, '.npmrc'), `//registry.npmjs.org/:_authToken=${process.env.NPM_TOKEN}`)
-
-spawn(npm, ['publish', ...args, ...tagArgs], {
-  stdio: 'inherit',
-  cwd: join(process.cwd(), targetPath)
-}).on('exit', code => {
-  process.exitCode = code
-}).on('error', e => console.error(e))
+  if (!semverValid(version)) {
+    console.error(`Error: invalid version "${version}"`)
+    process.exit(1)
+  }
+  
+  const prerelease = semverPrerelease(version)
+  const tag = (prerelease && prerelease.length) ? prerelease.filter(part => typeof part === 'string').join('.') || 'prerelease' : false
+  const tagArgs = tag ? ['--tag', tag] : []
+  const npm = /^win/.test(process.platform) ? 'npm.cmd' : 'npm'
+  
+  fs.writeFileSync(join(process.cwd(), targetPath, '.npmrc'), `//registry.npmjs.org/:_authToken=${process.env.NPM_TOKEN}`)
+  
+  spawn(npm, ['publish', ...args, ...tagArgs], {
+    stdio: 'inherit',
+    cwd: join(process.cwd(), targetPath)
+  }).on('exit', code => {
+    process.exitCode = code
+  }).on('error', e => console.error(e))
+})
