+++ tsconfig.json
@@ -24,7 +24,6 @@
     ]
   },
   "include": [
-    "./src/**/*.ts",
     "./types/**/*.d.ts"
   ],
   "exclude": [
